% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miRNASeq1}
\alias{miRNASeq1}
\title{miRNA Sequence dataset from TCGA project}
\format{
A data frame with 18 columns and 10 rows:
\describe{
\item{miRNA1}{RPM value for miRNA1 in each sample}
\item{miRNA2}{RPM value for miRNA2 in each sample}
...
\item{miRNA18}{RPM value for miRNA18 in each sample}
}
}
\source{
\url{https://bioconductor.org/packages/RTCGA.miRNASeq/} (RTCGA.miRNASeq package)
}
\usage{
miRNASeq1
}
\description{
A small subset of the BRCA.miRNASeq dataset from the RTCGA.miRNASeq package.
This dataset contains expression information for 10 samples and 18 miRNAs.
Each value in the data frame represents \strong{reads per million miRNA mapped (RPM)} for that sample and miRNA.
It is provided as an example dataset for demonstration and testing purposes.
}
\examples{
\dontrun{
miRNASeq1
}
}
\keyword{datasets}
