% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runmiRNARefine.R
\name{runmiRNARefine}
\alias{runmiRNARefine}
\title{Launch Shiny App for miRNARefine}
\usage{
runmiRNARefine()
}
\value{
No return value but open up a Shiny page.
}
\description{
A function that launches the Shiny app for miRNARefine.
The shiny app enables users to upload their own miRNA expression datasets,
preprocess the data according to users' selection, calculate stability
metrics (e.g., CV), and visualize the results. The app highlights the most
and least stable miRNAs on the scatter plot.
}
\examples{
\dontrun{

miRNARefine::runmiRNARefine()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}
}
