% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectBatch.R
\name{detectBatch}
\alias{detectBatch}
\title{Detect and Correct Batch Effects in miRNA Data}
\usage{
detectBatch(miRNAdata, batch = NULL, correct = FALSE, report_summary = TRUE)
}
\arguments{
\item{miRNAdata}{A numeric data frame or matrix where rows are samples and columns are miRNAs.}

\item{batch}{A factor or character vector indicating the batch assignment for each sample. Default NULL.}

\item{correct}{Logical, whether to perform batch effect correction using ComBat. Default FALSE.}

\item{report_summary}{Logical, whether to print summary information of batch differences before (and after) correction. Default TRUE.}
}
\value{
A list containing:
\describe{
\item{batch_effects}{Data frame summarizing per-batch means and variances.}
\item{corrected_data}{Batch-corrected expression matrix if \code{correct = TRUE}}
}
}
\description{
Detects and optionally corrects batch effects in the dataset. Ensures that
technical or experimental batch differences do not confound downstream
analyses.
}
\examples{
# Example 1:
# Using miRNASeq data available with package
data(miRNASeq2)
filled2 <- missingValueHandling(miRNAdata = miRNASeq2,
                                method = "median",
                                report_summary = FALSE)
result <- detectBatch(miRNAdata = filled2, correct = TRUE,
                      report_summary = FALSE)
result$corrected_data
result$batch_effects

\dontrun{
# Example 2:
# Obtain an external sample miRNASeq dataset
# Example requires the RTCGA.miRNASeq package:
if (requireNamespace("RTCGA.miRNASeq", quietly = TRUE)) {
  library(RTCGA.miRNASeq)
  dim(ACC.miRNASeq) # 240 1048

  sample <- RTCGA.miRNASeq::ACC.miRNASeq[1:100, 1:20]
  result <- detectBatch(miRNAdata = sample, correct = TRUE,
                        batch = rep(1:2, each = 50)
                        report_summary = FALSE)
  result$corrected_data
  result$batch_effects
}}

}
\references{
Bolstad B (2025). preprocessCore: A collection of pre-processing functions.
R package version 1.72.0.
\href{https://bioconductor.org/packages/preprocessCore}{Link}.

#' Chodor, W. (2025). \emph{RTCGA.miRNASeq: miRNASeq datasets from The Cancer
Genome Atlas Project}. R package version 1.36.0.

#' Hastie, T., Tibshirani, R., & Friedman, J. (2009).
\emph{The Elements of Statistical Learning}, 2nd Edition. Springer.

Johnson, W. E., Li, C., & Rabinovic, A. (2007). Adjusting batch effects in
microarray expression data using empirical Bayes methods.
\emph{Biostatistics}, 8(1), 118–127.

Leek, J. T., et al. (2010). Tackling the widespread and critical impact
of batch effects in high-throughput data. \emph{Nature Reviews Genetics},
11(10), 733–739.
}
