% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miRNASeq1}
\alias{miRNASeq1}
\title{miRNA Sequence dataset from TCGA project}
\format{
A data frame with 10 rows and 18 columns:
\describe{
\item{Sample1}{Expression value for miRNA 1 in sample 1}
\item{Sample2}{Expression value for miRNA 2 in sample 1}
...
\item{Sample10}{Expression value for miRNA 18 in sample 1}
}
}
\source{
\url{https://bioconductor.org/packages/RTCGA.miRNASeq/} (RTCGA.miRNASeq package)
}
\usage{
miRNASeq1
}
\description{
A small subset of the BRCA.miRNASeq dataset from the RTCGA.miRNASeq package.
This dataset contains expression information for 10 samples and 18 miRNAs.
Each value in the data frame represents \strong{reads per million miRNA mapped (RPM)} for that sample and miRNA.
It is provided as an example dataset for demonstration and testing purposes.
}
\examples{
\dontrun{
miRNASeq1
}
}
\keyword{datasets}
