% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miRNASeq1}
\alias{miRNASeq1}
\title{miRNA Sequence dataset from TCGA project}
\format{
A data frame with 18 columns and 10 rows:
\describe{
\item{hsa-let-7a-1}{RPM value for the miRNA in each sample}
\item{hsa-let-7a-2}{RPM value for the miRNA in each sample}
\item{hsa-let-7a-3}{RPM value for the miRNA in each sample}
\item{hsa-let-7b}{RPM value for the miRNA in each sample}
\item{hsa-let-7c}{RPM value for the miRNA in each sample}
\item{hsa-let-7d}{RPM value for the miRNA in each sample}
\item{hsa-let-7e}{RPM value for the miRNA in each sample}
\item{hsa-let-7f-1}{RPM value for the miRNA in each sample}
\item{hsa-let-7f-2}{RPM value for the miRNA in each sample}
\item{hsa-let-7g}{RPM value for the miRNA in each sample}
\item{hsa-let-7i}{RPM value for the miRNA in each sample}
\item{hsa-mir-1-1}{RPM value for the miRNA in each sample}
\item{hsa-mir-1-2}{RPM value for the miRNA in each sample}
\item{hsa-mir-100}{RPM value for the miRNA in each sample}
\item{hsa-mir-101-1}{RPM value for the miRNA in each sample}
\item{hsa-mir-101-2}{RPM value for the miRNA in each sample}
\item{hsa-mir-103-1}{RPM value for the miRNA in each sample}
\item{hsa-mir-103-1-as}{RPM value for the miRNA in each sample}
}
}
\source{
\url{https://bioconductor.org/packages/RTCGA.miRNASeq/} (RTCGA.miRNASeq package)
}
\usage{
miRNASeq1
}
\description{
A small subset of the BRCA.miRNASeq dataset from the RTCGA.miRNASeq package.
This dataset contains expression information for 10 samples and 18 miRNAs.
Each value in the data frame represents \strong{reads per million miRNA mapped (RPM)} for that sample and miRNA.
It is provided as an example dataset for demonstration and testing purposes.
}
\examples{
\dontrun{
miRNASeq1
}
}
\keyword{datasets}
