% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miRNASeq2}
\alias{miRNASeq2}
\title{Toy miRNA Dataset}
\format{
A data frame with 20 columns and 30 rows:
\describe{
\item{miRNA1}{RPM value for miRNA1 in each sample}
\item{miRNA2}{RPM value for miRNA2 in each sample}
\item{miRNA3}{RPM value for miRNA3 in each sample}
\item{miRNA4}{RPM value for miRNA4 in each sample}
\item{miRNA5}{RPM value for miRNA5 in each sample}
\item{miRNA6}{RPM value for miRNA6 in each sample}
\item{miRNA7}{RPM value for miRNA7 in each sample}
\item{miRNA8}{RPM value for miRNA8 in each sample}
\item{miRNA9}{RPM value for miRNA9 in each sample}
\item{miRNA10}{RPM value for miRNA10 in each sample}
\item{miRNA11}{RPM value for miRNA11 in each sample}
\item{miRNA12}{RPM value for miRNA12 in each sample}
\item{miRNA13}{RPM value for miRNA13 in each sample}
\item{miRNA14}{RPM value for miRNA14 in each sample}
\item{miRNA15}{RPM value for miRNA15 in each sample}
\item{miRNA16}{RPM value for miRNA16 in each sample}
\item{miRNA17}{RPM value for miRNA17 in each sample}
\item{miRNA18}{RPM value for miRNA18 in each sample}
\item{miRNA19}{RPM value for miRNA19 in each sample}
\item{miRNA20}{RPM value for miRNA20 in each sample}
}
}
\usage{
miRNASeq2
}
\description{
A small toy miRNA dataset including missing values and outliers.
This dataset is designed for testing preprocessing and filtering functions.
Each row represents a sample, and each column represents a miRNA feature.
Values represent \strong{reads per million miRNA mapped (RPM)}.
Some values are intentionally set as \code{NA} or extreme outliers
to simulate noisy, unclean data.
}
\examples{
\dontrun{
miRNASeq2
}
}
\keyword{datasets}
