% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRNAStability.R
\name{miRNAStability}
\alias{miRNAStability}
\title{miRNA Stability Analysis}
\usage{
miRNAStability(
  miRNAdata,
  metrics = c("CV", "MAD"),
  num_top = 5,
  report_summary = TRUE
)
}
\arguments{
\item{miRNAdata}{A numeric matrix or data frame with miRNAs as columns and samples as rows.}

\item{metrics}{A character vector specifying which stability metrics to calculate.
Options: \code{c("CV", "MAD")}. Default is both.}

\item{num_top}{n most/least stable miRNAs to be selected. Default 5.}

\item{report_summary}{Logical, whether to print a short summary of the most/least stable miRNAs. Default TRUE.}
}
\value{
A list with:
\describe{
\item{stability_scores}{A data frame containing CV and/or MAD for each miRNA.}
\item{most_stable}{Names of the top 5 most stable miRNAs.}
\item{least_stable}{Names of the top 5 least stable miRNAs.}
}
}
\description{
Calculates feature-level stability metrics such as Coefficient of Variation
(CV) and Median Absolute Deviation (MAD) for each miRNA across samples.
Helps identify the most stable miRNAs for reliable downstream analyses.
}
\examples{
# Example 1:
# Using miRNASeq1 available with package
data(miRNASeq1)
result <- miRNAStability(miRNAdata = miRNASeq1, report_summary = FALSE)

result$most_stable
result$least_stable

#Example 2:
# Using miRNASeq2 available with package
data(miRNASeq2)
filled2 <- missingValueHandling(miRNAdata = miRNASeq2, method = "median",
                                report_summary = FALSE)
result2 <- miRNAStability(miRNAdata = filled2, metrics = "CV",
                          report_summary = FALSE)
result2$most_stable
result2$least_stable

\dontrun{
# Example 3:
# Obtain an external sample miRNASeq dataset
# Example requires the RTCGA.miRNASeq package:
if (requireNamespace("RTCGA.miRNASeq", quietly = TRUE)) {
  library(RTCGA.miRNASeq)
  dim(ACC.miRNASeq) # 240 1048

  sample <- RTCGA.miRNASeq::ACC.miRNASeq[1:100, 1:20]
  result <- miRNAStability(miRNAdata = sample, report_summary = FALSE)

  result3$most_stable
  result3$least_stable
}}

}
\references{
Bolstad B (2025). preprocessCore: A collection of pre-processing functions.
R package version 1.72.0.
\href{https://bioconductor.org/packages/preprocessCore}{Link}.

Chodor, W. (2025). \emph{RTCGA.miRNASeq: miRNASeq datasets from The Cancer
Genome Atlas Project}. R package version 1.36.0.

Hastie, T., Tibshirani, R., & Friedman, J. (2009).
\emph{The Elements of Statistical Learning}, 2nd Edition. Springer.

Hima Bindu, A., Suresh, P., Reddy, R., et al. (2019). Systematic evaluation
of miRNA stability across biological replicates. \emph{Frontiers in Molecular
Biosciences}, 6, 153.
}
