% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataQuality.R
\name{missingValueHandling}
\alias{missingValueHandling}
\title{Handle Missing Values in miRNA Dataset}
\usage{
missingValueHandling(
  miRNAdata,
  method = c("median", "mean", "knn"),
  k = 5,
  report_summary = TRUE
)
}
\arguments{
\item{miRNAdata}{A numeric data frame or matrix, samples in rows, miRNAs in columns.}

\item{method}{Method for imputation: "mean", "median", or "knn". Default "median".}

\item{k}{Number of neighbors for KNN imputation (if method = "knn"). Default 5.}

\item{report_summary}{Logical, whether to print a summary of missing values before and after imputation. Default TRUE.}
}
\value{
A data frame or matrix with missing values imputed.
}
\description{
Detects and imputes missing values in a miRNA dataset. Supports adaptive
methods like mean, median, or KNN imputation (Campesato, 2023), ensuring
completeness for downstream analysis.
}
\examples{
# Example 1:
# Using miRNASeq data available with package
data(miRNASeq2)
filled <- missingValueHandling(miRNAdata = miRNASeq2,
                               method = "median",
                               k = 5,
                               report_summary = TRUE)
head(filled)

# Example 2:
# Obtain an external sample miRNASeq dataset
# Example requires the RTCGA.miRNASeq package:
\dontrun{
if (requireNamespace("RTCGA.miRNASeq", quietly = TRUE)) {
  library(RTCGA.miRNASeq)
  dim(ACC.miRNASeq) # 240 1048

  subset <- RTCGA.miRNASeq::ACC.miRNASeq[1:200, 1:20]
  sample <- subset[subset$miRNA_ID == "read_count", ]
  filled <- missingValueHandling(miRNAdata = sample,
                                 method = "knn",
                                 k = 5,
                                 report_summary = TRUE)
  head(filtered)
}}

}
\references{
BioConductor Project (2025). \emph{impute: Imputation for microarray data.}
\href{https://bioconductor.org/packages/impute/}{Link}.

Bolstad B (2025). preprocessCore: A collection of pre-processing functions.
R package version 1.72.0.
\href{https://bioconductor.org/packages/preprocessCore}{Link}.

Campesato, L. F. (2023). Adaptive missing value imputation methods for omics
data. \emph{Bioinformatics}, 39(7), btab789.

Chodor, W. (2025). \emph{RTCGA.miRNASeq: miRNASeq datasets from The Cancer
Genome Atlas Project}. R package version 1.36.0.

Hastie, T., Tibshirani, R., & Friedman, J. (2009).
\emph{The Elements of Statistical Learning}, 2nd Edition. Springer.

Robinson, M. D., McCarthy, D. J. and Smyth, G. K. (2010). edgeR: a
Bioconductor package for differential expression analysis of digital gene
expression data. \emph{Bioinformatics}, 26, 139-140.
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2796818/}{Link}
}
